Encapsulate the Canvas' Graphics object
The Canvas model needs to redraw the CanvasView whenever it is changed,
but this doesn't need to happen all the time in the event handlers if
the user hasn't actually changed anything on the Canvas. Therefore, it
makes the most sense to handle this during changes of the Canvas model,
which means the Graphics object has to be encapsulated to catch these
changes since an event handler can't be added to the Graphics object.